+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+###12.07.2016 v1.0.3
+
+- Print command output if error occured
+- Always send slack notification if error occured during deployment
+- Add colored output
+
 ###12.07.2016 v1.0.2
 
 - Major refactoring, no breaking changes